+++ task/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Sort unused tags alphabetically in tag editing UI.
 - Text editing starts with current value, instead of blank.
+- Avoid OS notches overlap of tag editing UI.
 
 # 0.0.3